+++ packages/components/src/grid-table/demo/树形数据.preview.vue
@@ -1,0 +1,72 @@
+<template>
+  <el-switch v-model="virtual" inactive-text="虚拟列表" />
+
+  <MixteGridTable
+    :columns
+    :data="data.data?.data.list"
+    :loading="data.isLoading"
+    :virtual
+    style="height: 360px"
+  >
+    <template #cell-email="{ value }">
+      <a :href="`mailto:${value}`">{{ value }}</a>
+    </template>
+  </MixteGridTable>
+</template>
+
+<script lang="tsx" setup>
+  import type { ResponseData, ResponseListData, User } from '@/types';
+  import { MixteGridTable } from '@mixte/components/grid-table';
+  import { defineTableColumns } from '@mixte/components/grid-table/utils';
+  import { useRequestReactive } from '@mixte/use';
+  import axios from 'axios';
+  import { ElButton, ElImage, ElTag } from 'element-plus';
+  import '@mixte/components/dist/grid-table/css/index.scss';
+
+  const data = useRequestReactive(() => {
+    return axios.post<ResponseData<ResponseListData<User>>>('https://m1.apifoxmock.com/m1/4781098-4434938-default/list/user', { pageSize: 10, treeData: true });
+  }, {
+    immediate: true,
+  });
+
+  const columns = defineTableColumns<User>([
+    {
+      field: 'name',
+      title: '姓名',
+      fixed: 'left',
+      render: ({ value, record }) => `${value} (${record.nameEn})`,
+    },
+    {
+      field: 'avatar',
+      title: '头像',
+      align: 'center',
+      render: ({ value }) => {
+        return <ElImage class="size-8" src={value} preview-src-list={[value]} preview-teleported lazy />;
+      },
+    },
+    { field: 'age', title: '年龄', align: 'center' },
+    { field: 'gender', title: '性别', align: 'center' },
+    { field: 'email', title: '邮箱' },
+    { field: 'address', title: '地址' },
+    {
+      field: 'statusValue',
+      title: '状态',
+      align: 'center',
+      render: ({ value, record }) => <ElTag type={value === 1 ? 'success' : 'info'}>{record.status}</ElTag>,
+    },
+    {
+      field: 'action',
+      title: '操作',
+      fixed: 'right',
+      align: 'center',
+      render: () => (
+        <div>
+          <ElButton type="primary" size="small" link>编辑</ElButton>
+          <ElButton type="danger" size="small" link>删除</ElButton>
+        </div>
+      ),
+    },
+  ]);
+
+  const virtual = ref(false);
+</script>