+++ packages/components/src/grid-table/src/composables/useShared.ts
@@ -1,9 +1,14 @@
 import type { GridTableProps, GridTableSlots } from '@mixte/components/grid-table/types';
-import type { StyleValue } from 'vue';
+import type { ComputedRef, ModelRef, StyleValue } from 'vue';
 import { columnIsFixedLeft, columnIsFixedRight } from '@mixte/components/grid-table/utils';
 import { createInjectionState, useElementSize, useScroll } from '@vueuse/core';
 import { isNumeric } from 'mixte';
+import { createTreeMate } from 'treemate';
 import { computed, reactive, ref, toValue, watch } from 'vue';
+
+interface UseSharedStoreOptions {
+  expandedRowKeys: ModelRef<string[]>;
+}
 
 export const [
   useSharedStore,
@@ -11,11 +16,16 @@ export const [
 ] = createInjectionState((
   props: GridTableProps<any>,
   slots: GridTableSlots<any>,
+  options: UseSharedStoreOptions,
 ) => {
+  const { expandedRowKeys } = options;
+
   /** 表格行主键 */
-  const rowKey = computed(() => {
-    return toValue(props.rowKey) ?? 'id';
-  });
+  const rowKey = computed(() => toValue(props.rowKey) ?? 'id');
+  /** 树形数据子节点字段名 */
+  const childrenKey = computed(() => props.childrenKey ?? 'children');
+  /** 树形数据展开的列主键 */
+  const expandColumnKey = computed(() => props.expandColumnKey ?? (props.columns?.[0]?.field ?? ''));
 
   /** 预渲染的行数 */
   const overscan = computed(() => {
@@ -27,6 +37,12 @@ export const [
   const estimatedRowHeight = computed(() => {
     let estimatedHeight = props.estimatedRowHeight;
     return isNumeric(estimatedHeight) && (estimatedHeight = Number(estimatedHeight)) > 0 ? estimatedHeight : 50;
+  });
+
+  const { treeMate, expandedRowSet, displayedData, updateExpanded } = useTreeData(props, {
+    rowKey,
+    childrenKey,
+    expandedRowKeys,
   });
 
   /** 所有固定在左侧的列 */
@@ -80,9 +96,18 @@ export const [
     props,
     slots,
 
+    expandedRowKeys,
+
     rowKey,
+    childrenKey,
+    expandColumnKey,
     overscan,
     estimatedRowHeight,
+
+    treeMate,
+    expandedRowSet,
+    displayedData,
+    updateExpanded,
 
     fixedLeftColumns,
     fixedRightColumns,
@@ -98,3 +123,53 @@ export const [
     tableTheadSize,
   };
 });
+
+export function useTreeData(
+  props: GridTableProps<any>,
+  options: {
+    rowKey: ComputedRef<string>;
+    childrenKey: ComputedRef<string>;
+    expandedRowKeys: UseSharedStoreOptions['expandedRowKeys'];
+  },
+) {
+  const { rowKey, childrenKey, expandedRowKeys } = options;
+
+  const treeMate = computed(() => {
+    return createTreeMate<Record<string, any>>(props.data ?? [], {
+      ignoreEmptyChildren: true,
+      getKey: row => row[rowKey.value],
+      getChildren: row => row[childrenKey.value],
+      getIsGroup: () => false,
+    });
+  });
+
+  // todo
+  //  - props.expandedRowKeys 变更时, 需要重新计算
+  const expandedRowSet = ref(new Set(expandedRowKeys.value));
+
+  const displayedData = computed(() => {
+    // 直接使用 expandedRowKeys.value 不知为什么没有被收集依赖
+    // 当 expandedRowKeys.value 改变时, computed 不会重新计算, 等有空再研究
+    return treeMate.value.getFlattenedNodes([...expandedRowSet.value]);
+  });
+
+  function updateExpanded(key: string) {
+    const index = expandedRowKeys.value.indexOf(key);
+
+    if (index > -1) {
+      expandedRowKeys.value.splice(index, 1);
+      expandedRowSet.value.delete(key);
+    }
+    else {
+      expandedRowKeys.value.push(key);
+      expandedRowSet.value.add(key);
+    }
+  }
+
+  return {
+    treeMate,
+    expandedRowSet,
+    displayedData,
+    updateExpanded,
+  };
+}